refactor: move smartscanner to v2/backyard
