+++ server/src/main/kotlin/com/thoughtworks/archguard/change/application/DiffChangeService.kt
@@ -5,7 +5,7 @@ import com.thoughtworks.archguard.change.domain.DiffChangeRepo
 import com.thoughtworks.archguard.scanner.domain.scanner.javaext.bs.ScanContext
 import com.thoughtworks.archguard.scanner.domain.system.BuildTool
 import com.thoughtworks.archguard.scanner.infrastructure.command.InMemoryConsumer
-import com.thoughtworks.archguard.smartscanner.StranglerScannerExecutor
+import com.thoughtworks.archguard.v2.backyard.smartscanner.StranglerScannerExecutor
 import com.thoughtworks.archguard.system_info.domain.SystemInfo
 import org.archguard.scanner.core.context.AnalyserType
 import org.springframework.stereotype.Service
