+++ server/src/main/kotlin/com/thoughtworks/archguard/v2/backyard/smartscanner/ScannerClient.kt
@@ -1,4 +1,4 @@
-package com.thoughtworks.archguard.smartscanner
+package com.thoughtworks.archguard.v2.backyard.smartscanner
 
 interface ScannerClient {
     fun send(command: ScannerCommand)
